Add audio_receiver.py, a GUI for an audio receiver.
This script is a GUI for an audio receiver script, you input the local host IP and port that the audio streamer is listening on. It displays the connection status to the streamer along with the volume that starts at 0 dB and maxes out at 60 dB (on the display, not the actual stream).